* implemented the same document node retrieval functionality as I did for xml documents (doc.get<type>("path.to.traverse.to.leaf")), with some appropriate error checking (requested type must now actually match the type inside the json document!)  * migrated + updated the floor config to the new json format (I also did some slight adjustments to how the compute toolchain settings work: there is now a default toolchain entry used for all backends, but which entries can overriden by each individual backend)  * xml/libxml2 is now fully optional (disable with "./build.sh no-xml") and now disabled by default on windows/msvc -> updated README  * version bump to v0.2.0-d11, as this is a very breaking change

+++ compute/compute_base.hpp
@@ -48,7 +48,7 @@ public:
 	
 	//! initializes the compute context/object
 	virtual void init(const bool use_platform_devices = false,
-					  const uint32_t platform_index = ~0u,
+					  const uint64_t platform_index = ~0ull,
 					  const bool gl_sharing = false,
 					  const unordered_set<string> whitelist = {}) = 0;
