+++ compute/metal/metal_compute.hpp
@@ -34,7 +34,7 @@ public:
 	~metal_compute() override {}
 	
 	void init(const bool use_platform_devices = false,
-			  const uint32_t platform_index = ~0u,
+			  const uint64_t platform_index = ~0ull,
 			  const bool gl_sharing = false,
 			  const unordered_set<string> whitelist = {}) override;
